git-gui: fix load_current_branch when HEAD is a symlink
Git allows HEAD to be a symbolic link to a refname. So, for example HEAD
can be a symbolic link pointing to "refs/heads/master".

But git-gui does not account for that, and so when that happens, it
assumes HEAD to be detached instead, since it is directly reading
"refs/heads/master". The contents of that would be just a commit hash.
HEAD contains just a commit hash when it is detached.

Ideally, one would want to use 'git-symbolic-ref' to get the current
branch name. But in fc4e8da (git-gui: Internalize symbolic-ref HEAD
reading logic, 2007-05-30) reading HEAD directly was preferred over
'git-symbolic-ref' to reduce fork+exec per rescan, improving performance
on fork heavy systems like Windows.

Reported-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Pratyush Yadav <me@yadavpratyush.com>