Add: --nopushbutton commandline switch. This will remove only the "push" button from the main screen. See also issue https://github.com/prati0100/git-gui/issues/73
e.g.
git gui --nopushbutton
git gui citool --nopushbutton

Motivation:
When using Git Gui only for committing, it's very easy to mistakenly click the "push" button. Because the "push" button is visually very close to the "commit" button, it's only a few pixels away. This commit introduces a switch to hide the "push" button, so it can never be mistakenly clicked anymore.