Add screenshots to the README
The README is written in Markdown which supports images. Having a couple
of screenshots in the README (and thus on the project page on Github)
would help new users immediately understand the value of git-gui.

The screenshots are chosen to illustrate some of the main features.

The images are scaled down and compressed with `optipng -v -o7 *.png` to
ensure they don't waste space despite being binary files.

Signed-off-by: Otto Kekäläinen <otto@kekalainen.net>